fix(typing): mypy annotations for numpy comparisons + buf ndarrays
The numpy upper bound in this PR pulled a version whose type stubs flag
two pre-existing patterns on CI:

- complementary_filter.py:117 — ``displacement > max_expected`` where
  both sides are Python-level numpy scalars. Cast to float explicitly.
- camera_tracker.py:141,197 — ``buf = np.ndarray(...)`` without an
  annotation. Add an explicit ``buf: np.ndarray`` hint.

Pure typing cleanup; no runtime change.

+++ src/osc_tracking/complementary_filter.py
@@ -112,8 +112,10 @@ class ComplementaryFilter:
             camera_position is not None
             and not np.allclose(state.last_valid_position, 0.0)
         ):
-            displacement = np.linalg.norm(camera_position - state.position)
-            max_expected = max(np.linalg.norm(state.velocity) * dt * 3.0, 0.5)
+            displacement = float(np.linalg.norm(camera_position - state.position))
+            max_expected = float(
+                max(np.linalg.norm(state.velocity) * dt * 3.0, 0.5)
+            )
             if displacement > max_expected and confidence < 0.9:
                 camera_position = None  # Reject outlier
